Add support for 24 hour mode.
This adds support for 12/24 hour mode switching.
AsteroidOS uses/sets a DConf value to enable and detect 12 hour mode.
This is stored under the user settings, so make this daemon run under the user session instead of root.
mlite is used to read the DConf value.
Then use the Android property_set() function defined by libhybris to change the expected property in libmcutool.so.

Signed-off-by: Darrel Griët <dgriet@gmail.com>