Alter function signatures for callback
This alters the function signatures to add the codename of each watch as
another level of namespace and makes the functions void(int) to be
compatible directly with CLI11's callbacks.

Change function signature for use as callback

This matches the change to the header file in changing the function to a
void(int).

Restructure using SpiMsg type

This restructures the code to use a SpiMsg type (alias of
std::array<uint8_t, 7>) and modifies to simplify use of functions as
callbacks.  Functions not part of the interface are static.

Use more features of CLI11

This relies more heavily on the callback feature of CLI11 and simplifies
and clarifies the interface.

Restore missing return

If the file does not open, the function should return after reporting
this fact rather than trying to write to the file.

Signed-off-by: Ed Beroset <beroset@ieee.org>